enable building on 64-bit architectures
 * Include stdint.h for intptr_t types.
 * Check for more headers and data types in configure.ac.
 * Use a union of integer and string pointer to store config values, instead
   of type-punning string pointers to integers which won't work on
   64-bit systems anyway.
 * Silence some "incompatible pointer types" GCC warnings, as
   pointers to integer and union config_u are in fact
   interchangeable.
 * Fix memcpy() size parameters and pointer progression.
 * Fix BMP screenshots by changing some variables from ambigious long
   to corresponding int types.
 * Silence the last remaining "cast pointer from/to integer of
   different size" warnings.

+++ Source/d_main.c
@@ -1501,7 +1501,7 @@ void D_DoomMain(void)
       // killough 5/2/98: this takes a memory
       // address as an integer on the command line!
 
-      statcopy = (void*) atoi(myargv[p+1]);
+      statcopy = (void*)(intptr_t) atoi(myargv[p+1]);
       puts("External statistics registered.");
     }
