+++ Source/d_net.c
@@ -91,7 +91,7 @@ extern int  key_escape;                // phares
 //
 int NetbufferSize (void)
 {
-  return (int)&(((doomdata_t *)0)->cmds[netbuffer->numtics]); 
+  return (intptr_t)&(((doomdata_t *)0)->cmds[netbuffer->numtics]);
 }
 
 //
@@ -109,7 +109,7 @@ unsigned NetbufferChecksum (void)
 //    return 0;                 // byte order problems
 //#endif
 
-  l = (NetbufferSize () - (int)&(((doomdata_t *)0)->retransmitfrom))/4;
+  l = (NetbufferSize () - (intptr_t)&(((doomdata_t *)0)->retransmitfrom))/4;
   for (i=0 ; i<l ; i++)
     c += ((unsigned *)&netbuffer->retransmitfrom)[i] * (i+1);
